Scrolling & Category Selection Added/Fixed
This update is very sloppy how ever it does the job. It will more then
likely be recoded down the rode how ever its fine the way it currently
works. This allows selection of categories and viewing the lists as well
as adding a new glow system limiting this to 1.8+ and adds category
changing and scrolling between the items in the categories!